refactor(message): move MessageSubcomponentProperty to Message.types and remove constants file
