+++ chat/message/src/Message/CompactMessage.tsx
@@ -7,7 +7,6 @@ import { filterChildren, findChild } from '@leafygreen-ui/lib';
 import { BaseFontSize } from '@leafygreen-ui/tokens';
 import { Body } from '@leafygreen-ui/typography';
 
-import { MessageSubcomponentProperty } from '../constants';
 import {
   MessageContainer,
   Variant as MessageContainerVariant,
@@ -18,7 +17,10 @@ import {
   avatarContainerStyles,
   getContainerStyles,
 } from './CompactMessage.styles';
-import { type MessageProps } from './Message.types';
+import {
+  type MessageProps,
+  MessageSubcomponentProperty,
+} from './Message.types';
 
 export const CompactMessage = forwardRef<HTMLDivElement, MessageProps>(
   (
